Long,ShortFormatterの作成、それぞれのクラスが持つ情報を整理
