+++ 08.ls_object/file_list.rb
@@ -8,6 +8,21 @@ class FileList
   def initialize(params, pathname)
     pattern = pathname.join('*')
     dot_params = params[:dot_match] ? [pattern, File::FNM_DOTMATCH] : [pattern]
-    @paths = Dir.glob(*dot_params).sort
+    @paths = params[:reverse] ? Dir.glob(*dot_params).sort.reverse : Dir.glob(*dot_params).sort
+  end
+
+  def self.collect_details(paths)
+    paths.map do |file_path|
+      FileDetail.new(file_path)
+    end
+  end
+
+  def self.research_max_size(file_details)
+    [
+      file_details.map { |file_detail| file_detail.nlink.size }.max,
+      file_details.map { |file_detail| file_detail.user.size }.max,
+      file_details.map { |file_detail| file_detail.group.size }.max,
+      file_details.map { |file_detail| file_detail.size.size }.max
+    ]
   end
 end
