+++ 08.ls_object/formatter.rb
@@ -3,70 +3,7 @@
 require './file_list'
 
 class Formatter
-  def initialize(file_list, params)
-    @file_paths = params[:reverse] ? file_list.paths.reverse! : file_list.paths
-  end
-
-  def ls_long
-    file_details = @file_paths.map do |file_path|
-      FileDetail.new(file_path)
-    end
-    total = "total #{count_block}"
-    body = render_long_format_body(file_details)
-    [total, *body].join("\n")
-  end
-
-  def count_block
-    @file_paths.sum do |file_path|
-      File::Stat.new(file_path).blocks
-    end
-  end
-
-  def render_long_format_body(file_details)
-    max_sizes = [
-      file_details.map { |file_detail| file_detail.nlink.size }.max,
-      file_details.map { |file_detail| file_detail.user.size }.max,
-      file_details.map { |file_detail| file_detail.group.size }.max,
-      file_details.map { |file_detail| file_detail.size.size }.max
-    ]
-    file_details.map do |data|
-      format_row(data, *max_sizes)
-    end
-  end
-
-  def format_row(data, max_nlink, max_user, max_group, max_size)
-    [
-      data.type_and_mode,
-      "  #{data.nlink.rjust(max_nlink)}",
-      " #{data.user.ljust(max_user)}",
-      "  #{data.group.ljust(max_group)}",
-      "  #{data.size.rjust(max_size)}",
-      " #{data.mtime}",
-      " #{data.basename}"
-    ].join
-  end
-
-  def ls_short
-    column = 3.0
-    row_count = (@file_paths.count.to_f / column).ceil
-    transposed_file_paths = safe_transpose(@file_paths.each_slice(row_count).to_a)
-    format_table(transposed_file_paths)
-  end
-
-  def safe_transpose(nested_file_names)
-    nested_file_names[0].zip(*nested_file_names[1..])
-  end
-
-  def format_table(file_paths)
-    file_paths.map do |row_files|
-      render_short_format_row(row_files)
-    end.join("\n")
-  end
-
-  def render_short_format_row(row_files)
-    row_files.map do |file_path|
-      basename = file_path ? File.basename(file_path) : ''
-      basename.ljust(30)
-    end.join.rstrip
+  def initialize(file_paths)
+    @file_paths = file_paths
   end
 end
